backend/requirements.txt
# FastAPI core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6

# HTTP client
httpx==0.26.0

# Async support
anyio==4.2.0

# Utilities
python-dateutil==2.8.2
email-validator==2.1.0.post1

# AI/ML (optional)
openai==1.9.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
